// frontend/src/api/models/AuditLogListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AuditLogResponse } from './AuditLogResponse'
export type AuditLogListResponse = {
  items: Array<AuditLogResponse>
  total: number
  page: number
  limit: number
}